finalizing before PR
Two known features/bugs that will be seperated out into a seperate PR.

Starting the Jobs using the new "schedules", so a batch poller to pick those up and start them.

Can't edit dropzone, this also occurs in the other monitoring that uses this selector so the fix should fix for both.