bug fixes - after wizard changes (#916)
* Finish my account pages

Adjust login methods to properly return all data (roles, applications).

Introduce session grabbing and revoking methods, as well as make small bug fixes to those backend routes.

* Clean up console logs

* Visual updates and swap ip for browser in device info

* add ToDO

* Update AuthHeader to add interceptor for token refresh

Fix the token refresh on the backend to append tokens to the reponse and then catching them on the frontend to replace it if a refreshed one is created.

Add message on login screen if user is logged out from an expired token

* Reintroduce generating new refresh token and new retry if new token is detected.

* New App Wizard Introduced

This Commit introduces the new application wizard to register an owner.

To do so, it creates the new wizard component, an unathenticated route to check if an owner account exists in the system to determine when to load it.

Some changes were also made to the BackendReducer and Authentication redux to accomodate the loading of this information and methods to regisater the first owner.

I also reworked the app.js component so that it will be much easier in the future to manage security.

* Add comments

* bug fixes

* Fix cluster tour, and where cluster screens grab user info from

Adjust assets screens and cluster screens to get user from local storage, not redux, avoids errors and sets up for security in the future.

* more bug fixes

* filter applications in redux to only include those that user has access too

* Fix application selector bugs

* Replace header: authHeader() with headers: authHeader()

header: authHeader() is not correct, headers: authHeader is the correct implementation. All replacements have been made throughout the application to avoid log outs due to malformed headers.