Move Request Access to Auth instead of Instance settings
Request access makes more sense in the auth api endpoint, which sits above the CSRF token middleware.

It has protection against spamming so CSRF token validation is unecessary for this route, as it is just making an access request, not any state changing request with meaning.

Could not resolve why csrf token was being returned as invalid for users without access, so this is technically a workaround.